Use WPF through .NET Core 3
We could go back and make the initial image paths be computed relative to the assembly location, but let's just forge ahead with two changes in one commit.

Yes, we are trying out the "Microsoft.NET.Sdk.WindowsDesktop" SDK, i.e., WPF in .NET Core 3.  To make that work with this code we needed to change a few things.
* We needed somewhere nice to get DelegateCommand.  Let's go with Prism.
* We needed to switch to using DelegateCommand<T> because the Prism.Commands.DelegateCommand constructor only takes an Action, and a "method group" doesn't fit in one of those.
* We needed to move some assembly metadata out of AssemblyInfo so as to not duplicate it (and thereby not be able to build!).
* Finally, as noted, so as to not hard code paths to the background and palette images, we intermixed that change in the "switch to WPF" change.